// apps/client/src/app/core/swagger/model/environment.ts
/**
 * NgReports
 * NG Reports API
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {EnvironmentUrl} from './environmentUrl';
import {Report} from './report';

export interface Environment {
  environmentId: string;
  name: string;
  description: string;
  urls: Array<EnvironmentUrl>;
  reports: Array<Report>;
}