// apps/client/src/app/core/swagger/model/updateEnvironment.ts
/**
 * NgReports
 * NG Reports API
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {EnvironmentUrl} from './environmentUrl';

export interface UpdateEnvironment {
  /**
   * The unique ID of the project this environment belongs to.
   */
  projectId?: string;
  /**
   * A unique identifier for the environment.
   */
  environmentId?: string;
  /**
   * The name of the environment.
   */
  name?: string;
  /**
   * A description of the environment.
   */
  description?: string;
  /**
   * List of URLs associated with this environment.
   */
  urls?: Array<EnvironmentUrl>;
}